+++ src/Yasumi/data/translations/ascensionDay.php
@@ -15,5 +15,6 @@ return [
     'en_US' => 'Ascension Day',
     'nl_NL' => 'Hemelvaart',
     'nl_BE' => 'Hemelvaart',
-    'fr_FR' => 'Ascension'
+    'fr_FR' => 'Ascension',
+    'da_DK' => 'Kristi Himmelfartsdag'
 ];