+++ src/Yasumi/data/translations/maundyThursday.php
@@ -12,5 +12,6 @@
 
 // Translation for Maundy Thursday
 return [
-    'es_ES' => 'Jueves Santo'
+    'es_ES' => 'Jueves Santo',
+    'da_DK' => 'Skærtorsdag'
 ];